+++ src/server/message.rs
@@ -61,10 +61,10 @@ impl Plugin for ServerMessagePlugin {
             .build_state(app.world_mut())
             .build_system(receive);
 
-        let broadcast_receive_fn = (
+        let receive_shared_fn = (
             FilteredResourcesMutParamBuilder::new(|builder| {
-                for message in registry.iter_all_broadcast() {
-                    builder.add_write_by_id(message.broadcast_id());
+                for message in registry.iter_all_shared() {
+                    builder.add_write_by_id(message.shared_messages_id());
                 }
             }),
             ParamBuilder,
@@ -72,7 +72,7 @@ impl Plugin for ServerMessagePlugin {
             ParamBuilder,
         )
             .build_state(app.world_mut())
-            .build_system(broadcast_receive);
+            .build_system(receive_shared);
 
         let trigger_fn = (
             FilteredResourcesMutParamBuilder::new(|builder| {
@@ -86,17 +86,17 @@ impl Plugin for ServerMessagePlugin {
             .build_state(app.world_mut())
             .build_system(trigger);
 
-        let broadcast_trigger_fn = (
+        let trigger_shared_fn = (
             FilteredResourcesMutParamBuilder::new(|builder| {
-                for event in registry.iter_broadcast_events() {
-                    builder.add_write_by_id(event.message().broadcast_id());
+                for event in registry.iter_shared_events() {
+                    builder.add_write_by_id(event.message().shared_messages_id());
                 }
             }),
             ParamBuilder,
             ParamBuilder,
         )
             .build_state(app.world_mut())
-            .build_system(broadcast_trigger);
+            .build_system(trigger_shared);
 
         let send_locally_fn = (
             FilteredResourcesMutParamBuilder::new(|builder| {
@@ -118,10 +118,10 @@ impl Plugin for ServerMessagePlugin {
             .add_systems(
                 PreUpdate,
                 (
-                    (receive_fn, broadcast_receive_fn).run_if(in_state(ServerState::Running)),
+                    (receive_fn, receive_shared_fn).run_if(in_state(ServerState::Running)),
                     (
                         trigger_fn.run_if(in_state(ClientState::Disconnected)),
-                        broadcast_trigger_fn,
+                        trigger_shared_fn,
                     ),
                 )
                     .chain()
@@ -204,8 +204,8 @@ fn receive(
     }
 }
 
-fn broadcast_receive(
-    mut broadcasts: FilteredResourcesMut,
+fn receive_shared(
+    mut shared_messages: FilteredResourcesMut,
     mut server_messages: ResMut<ServerMessages>,
     type_registry: Res<AppTypeRegistry>,
     message_registry: Res<RemoteMessageRegistry>,
@@ -214,13 +214,13 @@ fn broadcast_receive(
         type_registry: &type_registry,
     };
 
-    for message in message_registry.iter_all_broadcast() {
-        let broadcasts = broadcasts
-            .get_mut_by_id(message.broadcast_id())
-            .expect("broadcast messages resource should be accessible");
+    for message in message_registry.iter_all_shared() {
+        let shared_messages = shared_messages
+            .get_mut_by_id(message.shared_messages_id())
+            .expect("shared messages resource should be accessible");
 
         // SAFETY: passed pointer was obtained using this message data.
-        unsafe { message.receive(&mut ctx, broadcasts.into_inner(), &mut server_messages) };
+        unsafe { message.receive(&mut ctx, shared_messages.into_inner(), &mut server_messages) };
     }
 }
 
@@ -238,17 +238,17 @@ fn trigger(
     }
 }
 
-fn broadcast_trigger(
-    mut broadcasts: FilteredResourcesMut,
+fn trigger_shared(
+    mut shared_messages: FilteredResourcesMut,
     mut commands: Commands,
     registry: Res<RemoteMessageRegistry>,
 ) {
-    for event in registry.iter_broadcast_events() {
-        let broadcasts = broadcasts
-            .get_mut_by_id(event.message().broadcast_id())
-            .expect("broadcast messages resource should be accessible");
+    for event in registry.iter_shared_events() {
+        let shared_messages = shared_messages
+            .get_mut_by_id(event.message().shared_messages_id())
+            .expect("shared messages resource should be accessible");
         // SAFETY: passed pointer was obtained using this event data.
-        unsafe { event.trigger(&mut commands, broadcasts.into_inner()) };
+        unsafe { event.trigger(&mut commands, shared_messages.into_inner()) };
     }
 }