+++ src/shared/message/broadcast_event.rs
@@ -5,7 +5,7 @@ use log::debug;
 use serde::{Serialize, de::DeserializeOwned};
 
 use super::{
-    broadcast_message::BroadcastMessage,
+    broadcast_message::SharedMessage,
     client_message,
     ctx::{ClientSendCtx, ServerReceiveCtx},
     message_fns::{DeserializeFn, MessageFns, SerializeFn},
@@ -13,47 +13,47 @@ use super::{
 };
 use crate::prelude::*;
 
-/// An extension trait for [`App`] for creating broadcast events.
+/// An extension trait for [`App`] for creating shared events.
 ///
-/// See also [`BroadcastMessageAppExt`] for messages, [`ClientEventAppExt`] for regular client events
+/// See also [`SharedMessageAppExt`] for messages, [`ClientEventAppExt`] for regular client events
 /// and [`ServerEventAppExt`] for server events.
-pub trait BroadcastEventAppExt {
-    /// Registers a remote event that can be triggered using [`BroadcastTriggerExt::broadcast_trigger`].
+pub trait SharedEventAppExt {
+    /// Registers a remote event that can be triggered using [`SharedTriggerExt::shared_trigger`].
     ///
-    /// After triggering `E` event on the client, [`Broadcast<E>`] event will be triggered locally
-    /// with [`Broadcaster::Local`] and on the server with [`Broadcaster::Remote`].
+    /// After triggering `E` event on the client, [`LocalOrRemote<E>`] event will be triggered locally
+    /// with [`Sender::Local`] and on the server with [`Sender::Remote`].
     ///
-    /// On a listen server, the event will be triggered locally with [`Broadcaster::Local`].
-    fn add_broadcast_event<E: Event + Serialize + DeserializeOwned>(
+    /// On a listen server, the event will be triggered locally with [`Sender::Local`].
+    fn add_shared_event<E: Event + Serialize + DeserializeOwned>(
         &mut self,
         channel: Channel,
     ) -> &mut Self {
-        self.add_broadcast_event_with(
+        self.add_shared_event_with(
             channel,
             client_message::default_serialize::<E>,
             client_message::default_deserialize::<E>,
         )
     }
 
-    /// Same as [`Self::add_broadcast_event`], but additionally maps client entities to server inside the event before sending.
+    /// Same as [`Self::add_shared_event`], but additionally maps client entities to server inside the event before sending.
     ///
     /// Always use it for events that contain entities. Entities must be annotated with `#[entities]`.
     /// For details, see [`Component::map_entities`].
-    fn add_mapped_broadcast_event<E: Event + Serialize + DeserializeOwned + MapEntities + Clone>(
+    fn add_mapped_shared_event<E: Event + Serialize + DeserializeOwned + MapEntities + Clone>(
         &mut self,
         channel: Channel,
     ) -> &mut Self {
-        self.add_broadcast_event_with(
+        self.add_shared_event_with(
             channel,
             client_message::default_serialize_mapped::<E>,
             client_message::default_deserialize::<E>,
         )
     }
 
-    /// Same as [`Self::add_broadcast_event`], but uses the specified functions for serialization and deserialization.
+    /// Same as [`Self::add_shared_event`], but uses the specified functions for serialization and deserialization.
     ///
-    /// See also [`BroadcastMessageAppExt::add_broadcast_message_with`].
-    fn add_broadcast_event_with<E: Event>(
+    /// See also [`SharedMessageAppExt::add_shared_message_with`].
+    fn add_shared_event_with<E: Event>(
         &mut self,
         channel: Channel,
         serialize: SerializeFn<ClientSendCtx, E>,
@@ -61,8 +61,8 @@ pub trait BroadcastEventAppExt {
     ) -> &mut Self;
 }
 
-impl BroadcastEventAppExt for App {
-    fn add_broadcast_event_with<E: Event>(
+impl SharedEventAppExt for App {
+    fn add_shared_event_with<E: Event>(
         &mut self,
         channel: Channel,
         serialize: SerializeFn<ClientSendCtx, E>,
@@ -70,67 +70,62 @@ impl BroadcastEventAppExt for App {
     ) -> &mut Self {
         self.world_mut()
             .resource_mut::<ProtocolHasher>()
-            .add_broadcast_event::<E>();
+            .add_shared_event::<E>();
 
-        let fns =
-            MessageFns::new(serialize, deserialize).with_convert::<BroadcastEventMessage<E>>();
-        let event = BroadcastEvent::new(self, channel, fns);
+        let fns = MessageFns::new(serialize, deserialize).with_convert::<SharedEventMessage<E>>();
+        let event = SharedEvent::new(self, channel, fns);
         let mut registry = self.world_mut().resource_mut::<RemoteMessageRegistry>();
-        registry.register_broadcast_event(event);
+        registry.register_shared_event(event);
 
         self
     }
 }
 
-/// Small abstraction on top of [`BroadcastMessage`] that stores a function to trigger them.
-pub(crate) struct BroadcastEvent {
+/// Small abstraction on top of [`SharedMessage`] that stores a function to trigger them.
+pub(crate) struct SharedEvent {
     type_id: TypeId,
-    message: BroadcastMessage,
+    message: SharedMessage,
     trigger: TriggerFn,
 }
 
-impl BroadcastEvent {
+impl SharedEvent {
     fn new<E: Event>(
         app: &mut App,
         channel: Channel,
-        fns: MessageFns<ClientSendCtx, ServerReceiveCtx, BroadcastEventMessage<E>, E>,
+        fns: MessageFns<ClientSendCtx, ServerReceiveCtx, SharedEventMessage<E>, E>,
     ) -> Self {
         Self {
             type_id: TypeId::of::<E>(),
-            message: BroadcastMessage::new(app, channel, fns),
+            message: SharedMessage::new(app, channel, fns),
             trigger: Self::trigger_typed::<E>,
         }
     }
 
-    /// Drains received [`Broadcast<BroadcastEventMessage<E>>`] messages and triggers them as [`Broadcast<E>`].
+    /// Drains received [`LocalOrRemote<SharedEventMessage<E>>`] messages and triggers them as [`LocalOrRemote<E>`].
     ///
     /// # Safety
     ///
-    /// The caller must ensure that `broadcasts` is [`Messages<Broadcast<BroadcastEventMessage<E>>>`]
+    /// The caller must ensure that `shared_messages` is [`Messages<LocalOrRemote<SharedEventMessage<E>>>`]
     /// and this instance was created for `E`.
-    pub(crate) unsafe fn trigger(&self, commands: &mut Commands, broadcasts: PtrMut) {
-        unsafe { (self.trigger)(commands, broadcasts) }
+    pub(crate) unsafe fn trigger(&self, commands: &mut Commands, shared_messages: PtrMut) {
+        unsafe { (self.trigger)(commands, shared_messages) }
     }
 
     /// Typed version of [`Self::trigger`].
     ///
     /// # Safety
     ///
-    /// The caller must ensure that `broadcasts` is [`Messages<Broadcast<BroadcastEventMessage<E>>>`].
-    unsafe fn trigger_typed<E: Event>(commands: &mut Commands, broadcasts: PtrMut) {
-        let broadcasts: &mut Messages<Broadcast<BroadcastEventMessage<E>>> =
-            unsafe { broadcasts.deref_mut() };
-        for Broadcast {
-            broadcaster,
-            message,
-        } in broadcasts.drain()
-        {
+    /// The caller must ensure that `shared_messages` is [`Messages<LocalOrRemote<SharedEventMessage<E>>>`].
+    unsafe fn trigger_typed<E: Event>(commands: &mut Commands, shared_messages: PtrMut) {
+        let shared_messages: &mut Messages<LocalOrRemote<SharedEventMessage<E>>> =
+            unsafe { shared_messages.deref_mut() };
+        for LocalOrRemote { sender, message } in shared_messages.drain() {
             debug!(
-                "triggering `{}` from `{broadcaster:?}`",
-                ShortName::of::<Broadcast<E>>()
+                "triggering `{}` from `{sender:?}`",
+                ShortName::of::<LocalOrRemote<E>>()
             );
-            commands.trigger(Broadcast {
-                broadcaster,
+            commands.trigger(LocalOrRemote {
+                sender,
                 message: message.event,
             });
         }
@@ -140,51 +135,51 @@ impl BroadcastEvent {
         self.type_id
     }
 
-    pub(crate) fn message(&self) -> &BroadcastMessage {
+    pub(crate) fn message(&self) -> &SharedMessage {
         &self.message
     }
 }
 
-/// Signature of broadcast trigger functions.
+/// Signature of shared event trigger functions.
 type TriggerFn = unsafe fn(&mut Commands, PtrMut);
 
-/// Extension trait for triggering broadcast events.
+/// Extension trait for triggering shared events.
 ///
-/// See also [`BroadcastEventAppExt`].
-pub trait BroadcastTriggerExt {
-    /// Like [`Commands::trigger`], but triggers [`Broadcast<E>`] locally and on the server.
-    fn broadcast_trigger(&mut self, event: impl Event);
+/// See also [`SharedEventAppExt`].
+pub trait SharedTriggerExt {
+    /// Like [`Commands::trigger`], but triggers [`LocalOrRemote<E>`] locally and on the server.
+    fn shared_trigger(&mut self, event: impl Event);
 }
 
-impl BroadcastTriggerExt for Commands<'_, '_> {
-    fn broadcast_trigger(&mut self, event: impl Event) {
-        self.write_message(BroadcastEventMessage { event });
+impl SharedTriggerExt for Commands<'_, '_> {
+    fn shared_trigger(&mut self, event: impl Event) {
+        self.write_message(SharedEventMessage { event });
     }
 }
 
-impl BroadcastTriggerExt for World {
-    fn broadcast_trigger(&mut self, event: impl Event) {
-        self.write_message(BroadcastEventMessage { event });
+impl SharedTriggerExt for World {
+    fn shared_trigger(&mut self, event: impl Event) {
+        self.write_message(SharedEventMessage { event });
     }
 }
 
-/// A message that used under the hood for broadcast events.
+/// A message that used under the hood for shared events.
 ///
 /// Events are implemented through messages in order to reuse their logic.
 /// So we send this message instead and, after receiving it, drain it to trigger regular events.
 /// This also allows us to avoid requiring [`Clone`] because events can't be drained.
 #[derive(Message)]
-struct BroadcastEventMessage<E> {
+struct SharedEventMessage<E> {
     event: E,
 }
 
-impl<E> From<E> for BroadcastEventMessage<E> {
+impl<E> From<E> for SharedEventMessage<E> {
     fn from(event: E) -> Self {
         Self { event }
     }
 }
 
-impl<E> AsRef<E> for BroadcastEventMessage<E> {
+impl<E> AsRef<E> for SharedEventMessage<E> {
     fn as_ref(&self) -> &E {
         &self.event
     }