+++ src/shared/message/registry.rs
@@ -3,9 +3,8 @@ use core::any::TypeId;
 use bevy::prelude::*;
 
 use super::{
-    broadcast_event::BroadcastEvent, broadcast_message::BroadcastMessage,
-    client_event::ClientEvent, client_message::ClientMessage, server_event::ServerEvent,
-    server_message::ServerMessage,
+    broadcast_event::SharedEvent, broadcast_message::SharedMessage, client_event::ClientEvent,
+    client_message::ClientMessage, server_event::ServerEvent, server_message::ServerMessage,
 };
 
 /// Registered server and client messages and events.
@@ -15,10 +14,10 @@ pub struct RemoteMessageRegistry {
     // but they are messages under the hood.
     server_messages: Vec<ServerMessage>,
     client_messages: Vec<ClientMessage>,
-    broadcast_messages: Vec<BroadcastMessage>,
+    shared_messages: Vec<SharedMessage>,
     server_events: Vec<ServerEvent>,
     client_events: Vec<ClientEvent>,
-    broadcast_events: Vec<BroadcastEvent>,
+    shared_events: Vec<SharedEvent>,
 }
 
 impl RemoteMessageRegistry {
@@ -30,8 +29,8 @@ impl RemoteMessageRegistry {
         self.client_messages.push(message);
     }
 
-    pub(super) fn register_broadcast_message(&mut self, message: BroadcastMessage) {
-        self.broadcast_messages.push(message);
+    pub(super) fn register_shared_message(&mut self, message: SharedMessage) {
+        self.shared_messages.push(message);
     }
 
     pub(super) fn register_server_event(&mut self, event: ServerEvent) {
@@ -42,8 +41,8 @@ impl RemoteMessageRegistry {
         self.client_events.push(event);
     }
 
-    pub(super) fn register_broadcast_event(&mut self, event: BroadcastEvent) {
-        self.broadcast_events.push(event);
+    pub(super) fn register_shared_event(&mut self, event: SharedEvent) {
+        self.shared_events.push(event);
     }
 
     pub(super) fn iter_server_messages_mut(&mut self) -> impl Iterator<Item = &mut ServerMessage> {
@@ -66,10 +65,10 @@ impl RemoteMessageRegistry {
             .chain(self.client_events.iter().map(|e| e.message()))
     }
 
-    pub(crate) fn iter_all_broadcast(&self) -> impl Iterator<Item = &BroadcastMessage> {
-        self.broadcast_messages
+    pub(crate) fn iter_all_shared(&self) -> impl Iterator<Item = &SharedMessage> {
+        self.shared_messages
             .iter()
-            .chain(self.broadcast_events.iter().map(|e| e.message()))
+            .chain(self.shared_events.iter().map(|e| e.message()))
     }
 
     pub(crate) fn iter_server_events(&self) -> impl Iterator<Item = &ServerEvent> {
@@ -80,8 +79,8 @@ impl RemoteMessageRegistry {
         self.client_events.iter()
     }
 
-    pub(crate) fn iter_broadcast_events(&self) -> impl Iterator<Item = &BroadcastEvent> {
-        self.broadcast_events.iter()
+    pub(crate) fn iter_shared_events(&self) -> impl Iterator<Item = &SharedEvent> {
+        self.shared_events.iter()
     }
 
     /// Returns registered channel ID for server message `M`.
@@ -124,21 +123,21 @@ impl RemoteMessageRegistry {
             .map(|e| e.message().channel_id())
     }
 
-    /// Returns registered channel ID for broadcast message `M`.
+    /// Returns registered channel ID for shared message `M`.
     ///
-    /// See also [`BroadcastMessageAppExt::add_broadcast_message`](super::broadcast_message::BroadcastMessageAppExt::add_broadcast_message).
-    pub fn broadcast_message_channel<M: Message>(&self) -> Option<usize> {
-        self.broadcast_messages
+    /// See also [`SharedMessageAppExt::add_shared_message`](super::broadcast_message::SharedMessageAppExt::add_shared_message).
+    pub fn shared_message_channel<M: Message>(&self) -> Option<usize> {
+        self.shared_messages
             .iter()
             .find(|m| m.type_id() == TypeId::of::<M>())
             .map(|m| m.channel_id())
     }
 
-    /// Returns registered channel ID for broadcast event `E`.
+    /// Returns registered channel ID for shared event `E`.
     ///
-    /// See also [`BroadcastEventAppExt::add_broadcast_event`](super::broadcast_event::BroadcastEventAppExt::add_broadcast_event).
-    pub fn broadcast_event_channel<E: Event>(&self) -> Option<usize> {
-        self.broadcast_events
+    /// See also [`SharedEventAppExt::add_shared_event`](super::broadcast_event::SharedEventAppExt::add_shared_event).
+    pub fn shared_event_channel<E: Event>(&self) -> Option<usize> {
+        self.shared_events
             .iter()
             .find(|e| e.type_id() == TypeId::of::<E>())
             .map(|e| e.message().channel_id())